+++ test/assignment_pattern_loc_start/input.ts
@@ -1,0 +1,1 @@
+function conditional(item: string = 'a') {}