+++ test/class_class_abstract_method_with_unrelated_export/input.ts
@@ -1,0 +1,7 @@
+abstract class Student {
+ abstract study(): void
+}
+
+const a = 12345;
+
+export { a }