+++ test/class_private_id_class_method_duplicate/input.ts
@@ -1,0 +1,4 @@
+class Student {
+ #study(): void {}
+ #study(): void {}
+}