+++ test/class_private_id_class_method_override/input.ts
@@ -1,0 +1,7 @@
+class Student {
+ #study(book: 'math'): void
+ #study(book: 'english'): void
+ #study(book: 'math' | 'english'): void {
+   console.log('Im studying')
+ }
+}