+++ test/decorators_class_abstract/input.ts
@@ -1,0 +1,2 @@
+@decorator
+abstract class ExampleClass {}