+++ test/decorators_class_abstract_export/input.ts
@@ -1,0 +1,2 @@
+@decorator
+export abstract class ExampleClass {}